refactor importing Attachment model lines
