+++ apps/app/src/server/service/file-uploader/gcs.ts
@@ -2,7 +2,7 @@ import { Storage } from '@google-cloud/storage';
 import { Response } from 'express';
 import urljoin from 'url-join';
 
-import { IAttachmentDocument } from '~/server/models/attachment';
+import type { IAttachmentDocument } from '~/server/models';
 import loggerFactory from '~/utils/logger';
 
 import { configManager } from '../config-manager';
